Generate Plots for Analysing Feature Cell Abundance

This function processes clustering results, plots each cluster, and overlays each cluster's convex hull. It is adaptable to any number of cell_cluster_id.

plotGradCAMFeatureCells(
  x,
  feature_cells,
  p_adjust_method = "BH",
  ncol = 3,
  min_cells = 10,
  scatter_plot = FALSE,
  title = "GradCAM Feature Cells",
  Timer_positive = TRUE,
  ylim = NULL
)

Arguments

p_adjust_method

A method for p-value adjustment in multiple testing using Mann Whitney. clusteringFeatureCells cen be used.

ncol

Number of columns in output figure panel.

min_cells

Numeric. The minimum nunmber of cells within a cluster to be analysed. The default is 10.

scatter_plot

Logical. If TRUE, scatter plot for Angle and Intensity is generated.

Timer_positive

Logical. Whether to remove Timer negative cells.

ylim

Optional. A numeric vector of the length 2 for specifying ylim.

res_tockyrf

A list object output from `TockyKmeansRF`, which has been further processed using `ClusteringFeatureCells`.

Examples

if (FALSE) { # \dontrun{
  data <- data.frame(Angle = runif(100), Intensity = runif(100))
  cell_cluster_id <- dbscan(data, eps = 0.1, minPts = 5)$cluster
  plotGradCAMFeatureCells(data, cell_cluster_id)
} # }