Export TockyCNN Image Data from TockyPrepData Object

Export CNN images data from a TockyPrepData object and prints diagnostic information about the image data structure.

Usage

writeImages(x, numpy = TRUE)

Arguments

x

A TockyPrepData object containing processed CNN images data. This object must have already undergone ImageConversion.

numpy

Logical. Whether to export your data in the numpy format. If FALSE, csv files are exported instead.

Value

The CNN images data stored in the object's TockyCNNimages slot. The returned value maintains the original structure from the slot, typically including:

Image arrays in the first element

Sample definitions in the second element

Variables used in the third element

Details

This function performs the following actions:

Checks if the TockyCNNimages slot contains data

Prints diagnostic information including:

Dimensions of the image arrays

Variables used in the analysis

Sample definitions

Examples

if (FALSE) { # \dontrun{
# After successful ImageConversion:
images_data <- getCNNimages(tocky_prep_object)
dim(images_data[[1]])  # Access image array dimensions
} # }